# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Build Governance Platform",
...
    "version": "0.1.0-bgp-remediation"
  },
  "paths": {
...
    "/orgs/{tenant_id}/drafts": {
      "get": {
        "tags": [
          "drafts"
        ],
        "summary": "List Drafts",
        "description": "Own drafts only, regardless of tenant role -- a draft is never\nvisible to anyone but the user who saved it (Blueprint Sec.5.2: 'Owner\nscoped'), including tenant administrators.",
        "operationId": "list_drafts_orgs__tenant_id__drafts_get",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Id"
            }
          },
          {
            "name": "bgp_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Bgp Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DraftOut"
                  },
                  "title": "Response List Drafts Orgs  Tenant Id  Drafts Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/orgs/{tenant_id}/drafts/{draft_key}": {
      "get": {
        "tags": [
          "drafts"
        ],
        "summary": "Get Draft",
        "operationId": "get_draft_orgs__tenant_id__drafts__draft_key__get",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Id"
            }
          },
          {
            "name": "draft_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Draft Key"
            }
          },
          {
            "name": "bgp_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Bgp Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DraftOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "drafts"
        ],
        "summary": "Save Draft",
        "description": "Upsert with optimistic concurrency, same base_revision pattern as\nevidence revisions (app/routers/projects.py) -- base_revision=0 creates\na new draft; any other value must match the current revision exactly,\nor this is a 409 (two tabs/devices editing the same draft, not a\nsilent last-write-wins).",
        "operationId": "save_draft_orgs__tenant_id__drafts__draft_key__put",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Id"
            }
          },
          {
            "name": "draft_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Draft Key"
            }
          },
          {
            "name": "bgp_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Bgp Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveDraftRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DraftOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "drafts"
        ],
        "summary": "Discard Draft",
        "operationId": "discard_draft_orgs__tenant_id__drafts__draft_key__delete",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Id"
            }
          },
          {
            "name": "draft_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Draft Key"
            }
          },
          {
            "name": "bgp_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Bgp Session"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DraftOut": {
        "properties": {
          "draft_key": {
            "type": "string",
            "title": "Draft Key"
          },
          "form_data": {
            "type": "object",
            "title": "Form Data"
          },
          "revision": {
            "type": "integer",
            "title": "Revision"
          }
        },
        "type": "object",
        "required": [
          "draft_key",
          "form_data",
          "revision"
        ],
        "title": "DraftOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SaveDraftRequest": {
        "properties": {
          "base_revision": {
            "type": "integer",
            "title": "Base Revision"
          },
          "form_data": {
            "type": "object",
            "title": "Form Data"
          }
        },
        "type": "object",
        "required": [
          "base_revision",
          "form_data"
        ],
        "title": "SaveDraftRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}